+++ GraphNeuralNetworks/docs/make_tutorials_literate.jl
@@ -2,5 +2,8 @@ using Literate
 
 ENV["DATADEPS_ALWAYS_ACCEPT"] = true
 
+Literate.markdown("src_tutorials/introductory_tutorials/gnn_intro.jl", 
+                  "src/tutorials/"; execute = true)
+
 Literate.markdown("src_tutorials/introductory_tutorials/temporal_graph_classification.jl", 
                   "src/tutorials/"; execute = true)
